+++ source/javascripts/pages/WorkflowsPage/components/Drawers/Drawers.tsx
@@ -27,6 +27,7 @@ const Drawers = ({ children }: PropsWithChildren) => {
     setWorkflowId,
     unmountDialog,
     isDialogMounted,
+    parentWorkflowId,
   } = useWorkflowsPageStore();
 
   const { addStep, createWorkflow, getUniqueStepIds, addChainedWorkflow } = useBitriseYmlStore((s) => ({
@@ -140,6 +141,7 @@ const Drawers = ({ children }: PropsWithChildren) => {
           size="lg"
           context="workflow"
           workflowId={workflowId}
+          parentWorkflowId={parentWorkflowId}
           onRename={handleRenameWorkflow}
           isOpen={isDialogOpen(WorkflowsPageDialogType.WORKFLOW_CONFIG)}
           onClose={closeDialog}
